+++ packages/contracts/contracts/test/invariant/InvariantsBase.sol
@@ -245,12 +245,21 @@ contract InvariantsBase is Test {
         address[] memory allActors = depositManagerHandler.ghost_AllActors();
 
         for (uint256 i = 0; i < allActors.length; i++) {
-            assertLe(
-                allActors[i].balance,
-                depositManagerHandler.ghost_totalSuppliedGasCompensationFor(
-                    allActors[i]
-                )
-            );
+            uint256 actorBalance = allActors[i].balance;
+            uint256 actorBalanceCap = depositManagerHandler
+                .ghost_totalSuppliedGasCompensationFor(allActors[i]);
+
+            // NOTE: This invariant kept failing even though I checked all actor balances via
+            // logs and only found balance == ghost var but never greater than. I suspect there
+            // is a bug somewhere in foundry that doesn't like assertLe(0, 0) so hardcoding a
+            // statement here.
+            if (actorBalanceCap == actorBalance) {
+                continue;
+            }
+
+            if (actorBalance > actorBalanceCap) {
+                revert("actor balance exceeds expected cap");
+            }
         }
     }